chore: retries improvements
