+++ .gitignore
@@ -1,5 +1,5 @@
 .vscode
 .env
 .idea
-.lock
+deno.lock
 *.sh
