+++ src/Services/GithubApiService.ts
@@ -12,13 +12,12 @@ import {
   queryUserPullRequest,
   queryUserRepository,
 } from "../Schemas/index.ts";
-import { soxa } from "../../deps.ts";
 import { Retry } from "../Helpers/Retry.ts";
-import { GithubError, QueryDefaultResponse } from "../Types/index.ts";
 import { CONSTANTS } from "../utils.ts";
 import { EServiceKindError } from "../Types/EServiceKindError.ts";
 import { ServiceError } from "../Types/ServiceError.ts";
 import { Logger } from "../Helpers/Logger.ts";
+import { requestGithubData } from "./request.ts";
 
 // Need to be here - Exporting from another file makes array of null
 export const TOKENS = [
@@ -89,27 +88,6 @@ export class GithubApiService extends GithubRepository {
     );
   }
 
-  private handleError(responseErrors: GithubError[]): ServiceError {
-    const errors = responseErrors ?? [];
-
-    const isRateLimitExceeded = errors.some((error) =>
-      error.type.includes(EServiceKindError.RATE_LIMIT) ||
-      error.message.includes("rate limit")
-    );
-
-    if (isRateLimitExceeded) {
-      throw new ServiceError(
-        "Rate limit exceeded",
-        EServiceKindError.RATE_LIMIT,
-      );
-    }
-
-    throw new ServiceError(
-      "unknown error",
-      EServiceKindError.NOT_FOUND,
-    );
-  }
-
   async executeQuery<T = unknown>(
     query: string,
     variables: { [key: string]: string },
@@ -120,20 +98,10 @@ export class GithubApiService extends GithubRepository {
         CONSTANTS.DEFAULT_GITHUB_RETRY_DELAY,
       );
       const response = await retry.fetch<Promise<T>>(async ({ attempt }) => {
-        return await soxa.post("", {}, {
-          data: { query: query, variables },
-          headers: {
-            Authorization: `bearer ${TOKENS[attempt]}`,
-          },
-        });
-      }) as QueryDefaultResponse<{ user: T }>;
+        return await requestGithubData(query, variables, TOKENS[attempt]);
+      });
 
-      if (response?.data?.errors) {
-        return this.handleError(response?.data?.errors);
-      }
-
-      return response?.data?.data?.user ??
-        new ServiceError("not found", EServiceKindError.NOT_FOUND);
+      return response;
     } catch (error) {
       if (error instanceof ServiceError) {
         Logger.error(error);
